restyle StepCreationButton; break out into component from container
